boards.d: Reformat ESP partition filesystem for the BeaglePlay board
The bootROM for some SoCs (e.g: AM625[0] used in the BeaglePlay board) are
not able to boot when the ESP partition vfat filesystem geometry does not
match the expected sectors/track and head numbers.

Because the Fedora images are generated in VMs using raw images, they have
a 63 sectors/track, 16 heads geometry but for MMC block devices, it should
be 32 sectors/track, 64 heads for the AM625 SoC bootROM to boot the board.

To make sure that the ESP vfat filesystem will have the correct geometry,
reformat the filesystem of that partition using the actual block device
(e.g: MMC) that is going to be used to boot the board.

Some SoCs support loading the firmware binaries from block device sectors,
and so the files can be stored as raw data in the gap between the MBR and
the first partition. But this option is not available for the BeaglePlay
bootROM. For this reason, the only possible option is to store in the ESP.

[0]: https://www.ti.com/lit/gpn/am625

Suggested-by: Erico Nunes <nunes.erico@gmail.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>