Update the supported boards script and list
Update the supported boards list and script to be more
up to date and better reflect the aarch64 platforms we
actively support.

Also clarify in the help what supported in this context
actually means as we at times get equiries assuming it
means the device is fully supported in Fedora so clarify
that the support in this context is purely about writing
out the early boot firmware.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>