Unconditionally umount /tmp/fw
The spi-flashing-disk tool unconditionally mounts the storage device first
partition in the /tmp/fw mount point. But it only attempts to unmount that
if the /usr/share/uboot/${TARGET} directory exists and the u-boot binaries
can be copied.

Later though, it attempts to remove /tmp/fw which fails if the partition
is still mounted:

  sudo spi-flashing-disk --media=/dev/sda --target=rockpro64-rk3399
  Error: Partition(s) on /dev/sda are being used.
  Error: You requested a partition from 32.0MB to 256MB (sectors 62500..500000).
  The closest location we can manage is 31.9MB to 32.1MB (sectors 62220..62729).
  mkfs.fat 4.2 (2021-01-31)
  mkfs.vfat: /dev/sda1 contains a mounted filesystem.
  = No U-Boot files found for rockpro64-rk3399.
  rmdir: failed to remove '/tmp/fw': Device or resource busy

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>